sensor-app/requirements.txt
sounddevice>=0.5.0
python-socketio>=5.10.0
python-dotenv>=1.0.0
numpy>=1.24.0